add edit statys
