+++ main.cpp
@@ -1,10 +1,11 @@
+#pragma comment(lib,"User32.lib")
 #include "mainwindow.h"
 #include <QApplication>
 #include "database.h"
 #include "imgwork.h"
-
 int main(int argc, char *argv[])
 {
+
     QApplication a(argc, argv);
     Database db;
     db.initDataBase();
@@ -124,6 +125,11 @@ int main(int argc, char *argv[])
     db.createNodeAdventure("Морепродукты","Тайланд","Владивосток-Тайланд","sghf",12,"машина-самолет","облачно","ikgufl");
     db.createNodeAdventure("Бонжур","Франция","Австрия-Франция","kj",11,"поезд-машина","смерч","lhk");
     db.createNodeAdventure("Мясо и пиво","Чехия","Москва-Чехия","fh",16,"самолет","сухие грозы","mzxh");*/
+
+
+
+
+
     MainWindow w;
     w.show();
     return a.exec();
