+++ mainfunction.cpp
@@ -1,13 +1,12 @@
 #include "mainwindow.h"
 #include "ui_mainwindow.h"
+#include "treejson.h"
+#include "jsonwork.cpp"
 #include <QKeyEvent>
-
 #include <QTreeView>
 #include <QTreeWidgetItem>
 #include <QAbstractItemModel>
-#include <QJsonDocument>
-#include <QJsonObject>
-#include <QJsonArray>
+
 
 void viewDataBase(Ui::MainWindow* _ui){
     /*QSqlQuery query;
@@ -60,30 +59,10 @@ void simulatedClik(){
 
 }
 
-QJsonObject parseJson(){
-    // QString val = R"_({"list":[{"categoryname":"одежда","stuff":[{"name":"футболка","status":"not_put"},{"name":"шорты","status":"not_put"},{"name":"джинсы","status":"not_put"},{"name":"кепка","status":"not_put"},{"name":"купальник","status":"not_put"}]},{"categoryname":"обувь","stuff":[{"name":"кроссовки","status":"not_put"}]}]})_";
-     QString val;
-      QFile file;
-      file.setFileName("test.json");
-      file.open(QIODevice::ReadOnly | QIODevice::Text);
-      val = file.readAll();
-      file.close();
-      //qWarning() << val;
-      QJsonDocument document = QJsonDocument::fromJson(val.toUtf8());
-    return document.object();;
-}
-
-QJsonObject parseJson(QString val){
-    // QString val = R"_({"list":[{"categoryname":"одежда","stuff":[{"name":"футболка","status":"not_put"},{"name":"шорты","status":"not_put"},{"name":"джинсы","status":"not_put"},{"name":"кепка","status":"not_put"},{"name":"купальник","status":"not_put"}]},{"categoryname":"обувь","stuff":[{"name":"кроссовки","status":"not_put"}]}]})_";
-    QJsonDocument document = QJsonDocument::fromJson(val.toUtf8());
-    return document.object();;
-}
-
 void initTreeView(Ui::MainWindow* _ui,QString val ){
    _ui->treeWidget->clear();
    _ui->treeWidget->setColumnCount(2);
 
-   QList<QTreeWidgetItem *> listitems;
    QJsonObject root ;
 
    if(val != ""){
@@ -91,46 +70,7 @@ void initTreeView(Ui::MainWindow* _ui,QString val ){
    }else{
       root = parseJson();
    }
+   QList<QTreeWidgetItem *> listitems = TreeJson::jsonInTreeView(root);
 
-    QJsonValue jv = root.value("list");
-           if(jv.isArray()){
-               QJsonArray ja = jv.toArray();
-               for(int i = 0; i < ja.count(); i++){
-                   QJsonObject subtree = ja.at(i).toObject();
-                   for(auto key : subtree.keys() ){
-                        qWarning() <<" >"<<key;
-                        QTreeWidgetItem *item = new QTreeWidgetItem();
-                        item->setText(0, key);
-                        item->setText(1, "status");
-                        QJsonValue jvChild = subtree.value(key);
-                         if(jvChild.isArray()){
-                        QJsonArray jaChild = jvChild.toArray();
-                        for(int i = 0; i < jaChild.count(); i++){
-                            QJsonObject subtreeChild = jaChild.at(i).toObject();
-                            QTreeWidgetItem *child = new QTreeWidgetItem();
-                            for(auto keyChild : subtreeChild.keys() ){
-                                child->setText(0,keyChild);
-                                QString val = subtreeChild.value(keyChild).toString();
-                                child->setText(1, val);
-                                if(val == "not_put"){
-                                    child->setBackgroundColor(0, Qt::red);
-                                    child->setBackgroundColor(1, Qt::red);
-                                }
-                                    if( val == "put"){
-                                    child->setBackgroundColor(0, Qt::green);
-                                    child->setBackgroundColor(1, Qt::green);
-                                    }
-                                    if( val == "sale"){
-                                        child->setBackgroundColor(0, Qt::yellow);
-                                        child->setBackgroundColor(1, Qt::yellow);
-                                    }
-                                item->addChild(child);
-                            }
-
-                        }}
-                         listitems.push_back(item);
-                   }
-               }
-           }
-           _ui->treeWidget->addTopLevelItems(listitems);
+    _ui->treeWidget->addTopLevelItems(listitems);
 }
