+++ mainwindow.ui
@@ -25,7 +25,7 @@
      <rect>
       <x>10</x>
       <y>10</y>
-      <width>561</width>
+      <width>511</width>
       <height>251</height>
      </rect>
     </property>
@@ -33,8 +33,8 @@
    <widget class="QPushButton" name="removeBtn">
     <property name="geometry">
      <rect>
-      <x>580</x>
-      <y>10</y>
+      <x>560</x>
+      <y>20</y>
       <width>80</width>
       <height>51</height>
      </rect>
@@ -46,8 +46,8 @@
    <widget class="QPushButton" name="updateBtn">
     <property name="geometry">
      <rect>
-      <x>580</x>
-      <y>130</y>
+      <x>560</x>
+      <y>140</y>
       <width>81</width>
       <height>51</height>
      </rect>
@@ -59,8 +59,8 @@
    <widget class="QPushButton" name="addBtn">
     <property name="geometry">
      <rect>
-      <x>580</x>
-      <y>70</y>
+      <x>560</x>
+      <y>80</y>
       <width>80</width>
       <height>51</height>
      </rect>
@@ -157,9 +157,9 @@
     <property name="geometry">
      <rect>
       <x>350</x>
-      <y>670</y>
+      <y>660</y>
       <width>301</width>
-      <height>21</height>
+      <height>31</height>
      </rect>
     </property>
     <property name="text">
@@ -169,10 +169,10 @@
    <widget class="QPushButton" name="pushButton_9">
     <property name="geometry">
      <rect>
-      <x>350</x>
-      <y>640</y>
+      <x>10</x>
+      <y>660</y>
       <width>301</width>
-      <height>21</height>
+      <height>31</height>
      </rect>
     </property>
     <property name="text">
@@ -183,9 +183,9 @@
     <property name="geometry">
      <rect>
       <x>350</x>
-      <y>330</y>
+      <y>340</y>
       <width>300</width>
-      <height>231</height>
+      <height>250</height>
      </rect>
     </property>
     <property name="maximumSize">
@@ -195,14 +195,17 @@
      </size>
     </property>
     <property name="text">
-     <string>img[]</string>
+     <string/>
+    </property>
+    <property name="alignment">
+     <set>Qt::AlignCenter</set>
     </property>
    </widget>
    <widget class="Line" name="line">
     <property name="geometry">
      <rect>
       <x>320</x>
-      <y>620</y>
+      <y>640</y>
       <width>351</width>
       <height>20</height>
      </rect>
@@ -215,9 +218,9 @@
     <property name="geometry">
      <rect>
       <x>310</x>
-      <y>290</y>
+      <y>310</y>
       <width>20</width>
-      <height>411</height>
+      <height>381</height>
      </rect>
     </property>
     <property name="orientation">
@@ -227,9 +230,9 @@
    <widget class="Line" name="line_3">
     <property name="geometry">
      <rect>
-      <x>0</x>
-      <y>280</y>
-      <width>681</width>
+      <x>20</x>
+      <y>290</y>
+      <width>641</width>
       <height>20</height>
      </rect>
     </property>
@@ -241,7 +244,7 @@
     <property name="geometry">
      <rect>
       <x>10</x>
-      <y>300</y>
+      <y>310</y>
       <width>301</width>
       <height>301</height>
      </rect>
@@ -256,39 +259,48 @@
     <property name="geometry">
      <rect>
       <x>350</x>
-      <y>300</y>
+      <y>310</y>
       <width>301</width>
       <height>21</height>
      </rect>
     </property>
     <property name="text">
-     <string/>
+     <string>Название Поездки</string>
+    </property>
+    <property name="alignment">
+     <set>Qt::AlignCenter</set>
     </property>
    </widget>
    <widget class="QLabel" name="routTravel">
     <property name="geometry">
      <rect>
       <x>330</x>
-      <y>570</y>
+      <y>590</y>
       <width>331</width>
       <height>51</height>
      </rect>
     </property>
     <property name="text">
-     <string/>
+     <string>Маршрут поездки</string>
+    </property>
+    <property name="textFormat">
+     <enum>Qt::AutoText</enum>
+    </property>
+    <property name="alignment">
+     <set>Qt::AlignCenter</set>
     </property>
    </widget>
-   <widget class="QPushButton" name="saveBtn">
+   <widget class="Line" name="line_4">
     <property name="geometry">
      <rect>
-      <x>10</x>
-      <y>660</y>
-      <width>301</width>
-      <height>31</height>
+      <x>0</x>
+      <y>640</y>
+      <width>321</width>
+      <height>20</height>
      </rect>
     </property>
-    <property name="text">
-     <string>Save</string>
+    <property name="orientation">
+     <enum>Qt::Horizontal</enum>
     </property>
    </widget>
   </widget>
